Consolidate duplicate type coercion logic into shared module
Extract `_coerce_env` from `_container.py` and `_coerce`/`_parse_bool`
from `_config/_binding.py` into a single `coerce()` function in the new
`_coercion` private module.  Both call sites now delegate to the shared
implementation, which supports str, int, float, bool, Path, and
list[str].

Closes #112
